backend/requirements.txt
# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
setuptools==75.1.0

# --- Data Validation ---
pydantic==2.8.0
pydantic-settings==2.4.0
email-validator==2.2.0

# --- Database ---
sqlalchemy[asyncio]==2.0.35
asyncpg>=0.29.0
alembic==1.13.2

# --- Cache ---
redis[hiredis]==5.0.8

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# --- Market Data ---
yfinance==0.2.43
pandas>=2.2.2
numpy>=1.26.4
pandas-ta==0.4.71b0

# --- Task Queue ---
celery==5.4.0
flower==2.0.1

# --- NLP / Sentiment ---
textblob==0.18.0
requests==2.32.3
newsapi-python==0.2.7

# --- Rate Limiting ---
slowapi==0.1.9

# --- HTTP ---
httpx==0.27.0

# --- Utilities ---
python-dotenv==1.0.1
tenacity==8.5.0
structlog==24.4.0

# --- Testing ---
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
factory-boy==3.3.1
